requirements.txt
pandas==2.2.2
streamlit==1.37.0
openai==1.43.0
instructor==1.3.7
pydantic==2.8.2
pyperclip==1.9.0
azure-ai-documentintelligence==1.0.0b3
SQLAlchemy==2.0.31
psycopg2==2.9.9
